Only run evaluation if the output file doesn't exists

+++ code/13-sampling_robustness.R
@@ -44,25 +44,30 @@ scale <- F
 weight.type <- "asymmetry"
 
 1:length(onet) %>%
-	plyr::mlply(function(x){
+	plyr::m_ply(function(x){
 		
-		set.seed(this_task$replicate * this_task$completeness * 100)
-		n_edges_to_remove <- round(length(igraph::E(onet[[x]])) * (1 - this_task$completeness[1]))
-		edges_to_remove <- sample(igraph::E(onet[[x]]), n_edges_to_remove, prob = 1/igraph::E(onet[[x]])$weight)
-		subsampled_network <- igraph::delete_edges(onet[[x]], edges_to_remove)
+		message("Network ", names(onet)[x])
 		
-		o <- matched_frequency(subsampled_network, 
-													 prop = seq(0, 1, by = 0.1),
-													 weight.type = as.character(weight.type),
-													 scale = scale, 
-													 tmpdir = "/data/efc29/tmp/") 
-		saveRDS(o, 
-						file = paste0("./data/processed", 
-													"/sampling_robustness/",
-													weight.type, "/", "scaled_", scale, "/",
-													names(onet)[x], 
-													"_sampling_", this_task$completeness, 
-													"_replicate_", this_task$replicate, ".rds"), 
-						ascii = TRUE, compress = F)
-		return(o)
+		file_name <- paste0("./data/processed", 
+												"/sampling_robustness/",
+												weight.type, "/", "scaled_", scale, "/",
+												names(onet)[x], 
+												"_sampling_", this_task$completeness, 
+												"_replicate_", this_task$replicate, ".rds")
+		
+		if(!file.exists(file_name)){
+			set.seed(this_task$replicate * this_task$completeness * 100)
+			n_edges_to_remove <- round(length(igraph::E(onet[[x]])) * (1 - this_task$completeness[1]))
+			edges_to_remove <- sample(igraph::E(onet[[x]]), n_edges_to_remove, prob = 1/igraph::E(onet[[x]])$weight)
+			subsampled_network <- igraph::delete_edges(onet[[x]], edges_to_remove)
+			
+			o <- matched_frequency(subsampled_network, 
+														 prop = seq(0, 1, by = 0.1),
+														 weight.type = as.character(weight.type),
+														 scale = scale, 
+														 tmpdir = "/data/efc29/tmp/") 
+			saveRDS(o, 
+							file = , 
+							ascii = TRUE, compress = F)
+		}
 	}, .progress = "text")
